Add settings and Google Drive back up and import
Problems to fix:

Need to fix file paths to not be hard coded. Left it as this because getDatabasePath was not working.

importDatabaseFiles also does not fully close the database nor deletes the existing files.

startForResult functions are redundant for both back up and import. Made two separate ones because I didn't think to make drive a class variable until last minute.